feat: separates db/upsert-msr and check-eligibility routes logics into separate functions
