+++ src/app/check-eligibility/route.ts
@@ -1,121 +1,10 @@
-import * as Yup from "yup";
-import {
-	db,
-	getErrorMessage,
-	statusOnGoingMatch,
-	statusSupportRequestisAlreadyInQueue,
-	statusSupportRequestOngoingSocialWorker,
-} from "../../lib";
-import { MSRPiiSec, SupportType } from "@prisma/client";
-
-type MsrSearchResponse = {
-	supportRequestId: number | null;
-	zendeskTicketId: bigint | null;
-	shouldCreateMatch: boolean;
-};
-
-const payloadSchema = Yup.object({
-	email: Yup.string().email().required(),
-	supportType: Yup.string().oneOf(Object.values(SupportType)).required(),
-}).required();
-
-async function checkMatchEligibility(
-	supportType: SupportType,
-	msrId?: MSRPiiSec["msrId"]
-) {
-	if (!msrId) {
-		return {
-			supportRequestId: null,
-			zendeskTicketId: null,
-			shouldCreateMatch: true,
-		};
-	}
-
-	const ongoingMatch = await db.matches.findFirst({
-		where: {
-			msrId: msrId,
-			status: { in: statusOnGoingMatch },
-			supportType: supportType,
-		},
-		orderBy: {
-			createdAt: "desc",
-		},
-		select: {
-			supportRequestId: true,
-			status: true,
-			supportType: true,
-			msrZendeskTicketId: true,
-		},
-	});
-
-	if (ongoingMatch) {
-		return {
-			supportRequestId: ongoingMatch.supportRequestId,
-			zendeskTicketId: ongoingMatch.msrZendeskTicketId,
-			shouldCreateMatch: false,
-		};
-	}
-
-	const supportRequest = await db.supportRequests.findFirst({
-		where: {
-			msrId: msrId,
-			supportType,
-			status: { not: "duplicated" },
-		},
-		orderBy: {
-			createdAt: "asc",
-		},
-		select: {
-			supportRequestId: true,
-			status: true,
-			supportType: true,
-			zendeskTicketId: true,
-		},
-	});
-
-	if (!supportRequest) {
-		return {
-			supportRequestId: null,
-			zendeskTicketId: null,
-			shouldCreateMatch: true,
-		};
-	}
-
-	const statusOngoingSupportRequest: string[] = [
-		...statusSupportRequestisAlreadyInQueue,
-		...statusSupportRequestOngoingSocialWorker,
-	];
-
-	const hasOngoingSupport = statusOngoingSupportRequest.includes(
-		supportRequest.status
-	);
-
-	return {
-		supportRequestId: supportRequest.supportRequestId,
-		zendeskTicketId: supportRequest.zendeskTicketId,
-		shouldCreateMatch: !hasOngoingSupport,
-	};
-}
+import { getErrorMessage, checkMatchEligibility } from "../../lib";
 
 export async function POST(request: Request) {
 	try {
 		const payload = await request.json();
-		await payloadSchema.validate(payload);
 
-		const msr = await db.mSRPiiSec.findUnique({
-			where: {
-				email: payload.email,
-			},
-			select: {
-				msrId: true,
-				email: true,
-			},
-		});
-		const supportType: SupportType = payload.supportType;
-		const matchEligibilityResponse: MsrSearchResponse =
-			await checkMatchEligibility(supportType, msr?.msrId);
-
-		return Response.json(matchEligibilityResponse);
+		return await checkMatchEligibility(payload);
 	} catch (e) {
 		const error = e as Record<string, unknown>;
 		if (error["name"] === "ValidationError") {
